chore: Rename Googl-auth to Google-auth and update import references
