+++ src/presentation/pages/public/components/index.ts
@@ -1,1 +1,1 @@
-export * from './Googl-auth';
+export * from './Google-auth';
